requirements.txt
Django==1.10.3
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
